Fix: silence -Wsign-compare in dkim_dstring_cat / dkim_dstring_catn
struct dkim_dstring keeps ds_max/ds_alloc as int but the local
'needed' total is size_t.  Cast the int field to size_t at the two
comparisons in each function; the ds_max > 0 guard rules out the
negative-value case for the cap check, and ds_alloc is always a
non-negative buffer size.

+++ libopendkim/dkim-util.c
@@ -410,11 +410,11 @@ dkim_dstring_cat(struct dkim_dstring *dstr, unsigned char *str)
 	needed = dstr->ds_len + len;
 
 	/* too big? */
-	if (dstr->ds_max > 0 && needed >= dstr->ds_max)
+	if (dstr->ds_max > 0 && needed >= (size_t) dstr->ds_max)
 		return FALSE;
 
 	/* fits now? */
-	if (dstr->ds_alloc <= needed)
+	if ((size_t) dstr->ds_alloc <= needed)
 	{
 		/* nope; try to resize */
 		if (!dkim_dstring_resize(dstr, needed + 1))
@@ -496,11 +496,11 @@ dkim_dstring_catn(struct dkim_dstring *dstr, unsigned char *str, size_t nbytes)
 	needed = dstr->ds_len + nbytes;
 
 	/* too big? */
-	if (dstr->ds_max > 0 && needed >= dstr->ds_max)
+	if (dstr->ds_max > 0 && needed >= (size_t) dstr->ds_max)
 		return FALSE;
 
 	/* fits now? */
-	if (dstr->ds_alloc <= needed)
+	if ((size_t) dstr->ds_alloc <= needed)
 	{
 		/* nope; try to resize */
 		if (!dkim_dstring_resize(dstr, needed + 1))
